+++ lib/dimigoin_flutter_plugin.dart
@@ -6,10 +6,14 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 part 'services/dimigoin_login.dart';
 part 'services/dimigoin_meal.dart';
+part 'services/dalgeurak_service.dart';
 
 
 final _dio = Dio();
 final _storage = const FlutterSecureStorage();
+
+const apiUrl = "https://api.dimigo.in";
+
 
 DimigoinLogin _dimigoinLogin = DimigoinLogin();