+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- **ADD**: Dalgeurak Service API
+- **REFACTOR**: SourceCode Structure
+
 ## 0.1.2
 
 - **ADD**: Dimigoin Account Login - validateAccessToken, checkNowLogin Function